Use "unsigned char" instead of "char" when comparing byte sequences
The original bug was found when developing Juila parser in u-ctags
repo[1].  When searching a tag name "α", readtags reported nothing
though an entry for "α" was in the tags file used for testing
the Julia parser.

"α" is {206, 177} as a byte sequence.

However, when comparing bytes in strings, libreadtags used (singed)
char type. In the case "α" was {-50, -79}.  Using {-50, -79} for
comparison, the binary search didn't work as we expected.

The test for this change will be done in u-ctag side.

[1] https://github.com/universal-ctags/ctags/pull/2654

Signed-off-by: Masatake YAMATO <yamato@redhat.com>