[BREAKING API] add "tag" as prefix for "sortType"
The type name 'sortType' has no library specific prefix.
This can make hard to use libreadtags in an application because
of a type name cofliction.

To avoid the potential name confliction, this change renames
"sortType" to "tagSortType".

To keep application code using the name sortType work, "sortType"
is defined as a macro that can be used an alternative name to
"tagSortType".

The "sortType" macro is not defined if an application code defines
TAG_NO_COMPAT_SORT_TYPE.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>