Exemplify the relation between LT_VERSION and the string that can be passed -release option of libtool
We have used Autotools for building libreadtags.so.  To support both
Autotools and Cmake, we must clarify what the maintainer has to do
when releasing a new version.

An issue supporting multiple build-sys is the consistency of the
interface version of libreadtags.so.

If we use only Autotools, what the maintainer has to do is just
update LT_VERSION in configure.ac. libtool calculates the string
appended to libreadtags.so.<>. When building libreadtags.so, cmake
doesn't use libtool. To make the name of the library file the same
in both build systems, the maintainer may have to specify the
string calculated by libtool to BUILD_VERSION of CMakeLists.txt.

To avoid a potential mistake, I decided to record the strings to
configure.ac.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>